Abstract functionalities of spout implementations, implements #617
