+++ external/elasticsearch/es-conf.yaml
@@ -44,10 +44,10 @@ config:
   # es.status.filterQuery: "-(metadata.hostname:stormcrawler.net)"
 
   # time in secs for which the URLs will be considered for fetching after a ack of fail
-  es.status.ttl.purgatory: 30
+  spout.ttl.purgatory: 30
   
   # Min time (in msecs) to allow between 2 successive queries to ES
-  es.status.min.delay.queries: 2000
+  spout.min.delay.queries: 2000
 
   es.status.max.buckets: 50
   es.status.max.urls.per.bucket: 2
