+++ external/elasticsearch/src/main/java/com/digitalpebble/stormcrawler/elasticsearch/persistence/AbstractSpout.java
@@ -19,42 +19,24 @@ package com.digitalpebble.stormcrawler.elasticsearch.persistence;
 
 import java.io.IOException;
 import java.util.Date;
-import java.util.HashMap;
 import java.util.Iterator;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
-import java.util.Queue;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicBoolean;
 
-import org.apache.storm.metric.api.IMetric;
-import org.apache.storm.metric.api.MultiCountMetric;
 import org.apache.storm.spout.SpoutOutputCollector;
 import org.apache.storm.task.TopologyContext;
-import org.apache.storm.topology.OutputFieldsDeclarer;
-import org.apache.storm.topology.base.BaseRichSpout;
-import org.apache.storm.tuple.Fields;
-import org.apache.storm.tuple.Values;
-import org.apache.storm.utils.Utils;
-import org.elasticsearch.action.admin.cluster.shards.ClusterSearchShardsGroup;
-import org.elasticsearch.action.admin.cluster.shards.ClusterSearchShardsRequest;
-import org.elasticsearch.action.admin.cluster.shards.ClusterSearchShardsResponse;
-import org.elasticsearch.client.Client;
 import org.elasticsearch.client.RestHighLevelClient;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.digitalpebble.stormcrawler.Metadata;
 import com.digitalpebble.stormcrawler.elasticsearch.ElasticSearchConnection;
+import com.digitalpebble.stormcrawler.persistence.AbstractQueryingSpout;
 import com.digitalpebble.stormcrawler.util.CollectionMetric;
 import com.digitalpebble.stormcrawler.util.ConfUtils;
-import com.google.common.base.Optional;
-import com.google.common.cache.Cache;
-import com.google.common.cache.CacheBuilder;
 
-public abstract class AbstractSpout extends BaseRichSpout {
+public abstract class AbstractSpout extends AbstractQueryingSpout {
 
     private static final Logger LOG = LoggerFactory
             .getLogger(AbstractSpout.class);
@@ -66,6 +48,9 @@ public abstract class AbstractSpout extends BaseRichSpout {
     /**
      * Time in seconds for which acked or failed URLs will be considered for
      * fetching again, default 30 secs.
+     * 
+     * @deprecated replaced by
+     *             {@link com.digitalpebble.stormcrawler.persistence.AbstractQueryingSpout.StatusTTLPurgatory}
      **/
     protected static final String ESStatusTTLPurgatory = "es.status.ttl.purgatory";
 
@@ -88,6 +73,10 @@ public abstract class AbstractSpout extends BaseRichSpout {
     /**
      * Min time to allow between 2 successive queries to ES. Value in msecs,
      * default 2000.
+     * 
+     * @deprecated replaced by
+     *             {@link com.digitalpebble.stormcrawler.persistence.AbstractQueryingSpout.StatusMinDelayParamName}
+     * 
      **/
     private static final String ESStatusMinDelayParamName = "es.status.min.delay.queries";
 
@@ -99,9 +88,6 @@ public abstract class AbstractSpout extends BaseRichSpout {
 
     protected String indexName;
     protected String docType;
-    protected boolean active = true;
-    protected SpoutOutputCollector _collector;
-    protected MultiCountMetric eventCounter;
 
     protected static RestHighLevelClient client;
 
@@ -114,23 +100,6 @@ public abstract class AbstractSpout extends BaseRichSpout {
 
     /** Used to distinguish between instances in the logs **/
     protected String logIdprefix = "";
-
-    protected Queue<Values> buffer = new LinkedList<>();
-
-    /**
-     * Map to keep in-process URLs, ev. with additional information for URL /
-     * politeness bucket (hostname / domain etc.). The entries are kept in a
-     * cache for a configurable amount of time to avoid that some items are
-     * fetched a second time if new items are queried shortly after they have
-     * been acked.
-     */
-    protected InProcessMap<String, String> beingProcessed;
-
-    protected long timeStartESQuery = 0;
-
-    private long minDelayBetweenQueries = 2000;
-
-    protected AtomicBoolean isInESQuery = new AtomicBoolean(false);
 
     /** Field name used for field collapsing e.g. metadata.hostname **/
     protected String partitionField;
@@ -148,36 +117,6 @@ public abstract class AbstractSpout extends BaseRichSpout {
     protected Date lastDate;
 
     protected int resetFetchDateAfterNSecs = 120;
-
-    /** Map which holds elements some additional time after the removal. */
-    public class InProcessMap<K, V> extends HashMap<K, V> {
-
-        private Cache<K, Optional<V>> deletionCache;
-
-        public InProcessMap(long maxDuration, TimeUnit timeUnit) {
-            deletionCache = CacheBuilder.newBuilder()
-                    .expireAfterWrite(maxDuration, timeUnit).build();
-        }
-
-        @Override
-        public boolean containsKey(Object key) {
-            boolean incache = super.containsKey(key);
-            if (!incache) {
-                incache = (deletionCache.getIfPresent(key) != null);
-            }
-            return incache;
-        }
-
-        @Override
-        public V remove(Object key) {
-            deletionCache.put((K) key, Optional.absent());
-            return super.remove(key);
-        }
-
-        public long inCache() {
-            return deletionCache.size();
-        }
-    }
 
     @Override
     public void open(Map stormConf, TopologyContext context,
@@ -238,14 +177,6 @@ public abstract class AbstractSpout extends BaseRichSpout {
             LOG.info("{} assigned shard ID {}", logIdprefix, shardID);
         }
 
-        _collector = collector;
-
-        int ttlPurgatory = ConfUtils
-                .getInt(stormConf, ESStatusTTLPurgatory, 30);
-
-        minDelayBetweenQueries = ConfUtils.getLong(stormConf,
-                ESStatusMinDelayParamName, 2000);
-
         partitionField = ConfUtils.getString(stormConf,
                 ESStatusBucketFieldParamName, "metadata.hostname");
 
@@ -263,90 +194,10 @@ public abstract class AbstractSpout extends BaseRichSpout {
         resetFetchDateAfterNSecs = ConfUtils.getInt(stormConf,
                 ESStatusResetFetchDateParamName, resetFetchDateAfterNSecs);
 
-        beingProcessed = new InProcessMap<>(ttlPurgatory, TimeUnit.SECONDS);
-
         filterQuery = ConfUtils.getString(stormConf, ESStatusFilterParamName);
-
-        eventCounter = context.registerMetric("counters",
-                new MultiCountMetric(), 10);
-
-        context.registerMetric("buffer_size", new IMetric() {
-            @Override
-            public Object getValueAndReset() {
-                return buffer.size();
-            }
-        }, 10);
-
-        context.registerMetric("beingProcessed", new IMetric() {
-            @Override
-            public Object getValueAndReset() {
-                return beingProcessed.size();
-            }
-        }, 10);
-
-        context.registerMetric("inPurgatory", new IMetric() {
-            @Override
-            public Object getValueAndReset() {
-                return beingProcessed.inCache();
-            }
-        }, 10);
 
         esQueryTimes = new CollectionMetric();
         context.registerMetric("ES_query_time_msec", esQueryTimes, 10);
-    }
-
-    /** Returns true if ES was queried too recently and needs throttling **/
-    protected boolean throttleESQueries() {
-        Date now = new Date();
-        if (timeStartESQuery != 0) {
-            // check that we allowed some time between queries
-            long difference = now.getTime() - timeStartESQuery;
-            if (difference < minDelayBetweenQueries) {
-                long sleepTime = minDelayBetweenQueries - difference;
-                LOG.debug(
-                        "{} Not enough time elapsed since {} - should try again in {}",
-                        logIdprefix, timeStartESQuery, sleepTime);
-                return true;
-            }
-        }
-        return false;
-    }
-
-    @Override
-    public void declareOutputFields(OutputFieldsDeclarer declarer) {
-        declarer.declare(new Fields("url", "metadata"));
-    }
-
-    @Override
-    public void nextTuple() {
-
-        // inactive?
-        if (active == false)
-            return;
-
-        synchronized (buffer) {
-            // have anything in the buffer?
-            if (!buffer.isEmpty()) {
-                Values fields = buffer.remove();
-                String url = fields.get(0).toString();
-                beingProcessed.put(url, null);
-                _collector.emit(fields, url);
-                eventCounter.scope("emitted").incrBy(1);
-                return;
-            }
-        }
-
-        // check that we allowed some time between queries
-        // and not in middle of querying ES
-        if (isInESQuery.get() || throttleESQueries()) {
-            // sleep for a bit but not too much in order to give ack/fail a
-            // chance
-            Utils.sleep(10);
-            return;
-        }
-
-        // re-populate the buffer
-        populateBuffer();
     }
 
     /** Builds a query and use it retrieve the results from ES **/
@@ -381,25 +232,13 @@ public abstract class AbstractSpout extends BaseRichSpout {
     @Override
     public void ack(Object msgId) {
         LOG.debug("{}  Ack for {}", logIdprefix, msgId);
-        beingProcessed.remove(msgId);
-        eventCounter.scope("acked").incrBy(1);
+        super.ack(msgId);
     }
 
     @Override
     public void fail(Object msgId) {
         LOG.info("{}  Fail for {}", logIdprefix, msgId);
-        beingProcessed.remove(msgId);
-        eventCounter.scope("failed").incrBy(1);
-    }
-
-    @Override
-    public void activate() {
-        active = true;
-    }
-
-    @Override
-    public void deactivate() {
-        active = false;
+        super.fail(msgId);
     }
 
     @Override
