+++ external/solr/solr-conf.yaml
@@ -8,6 +8,13 @@ config:
   solr.status.bucket.field: host
   solr.status.bucket.maxsize: 100
   solr.status.metadata.prefix: metadata
+  solr.status.max.results: 
+  
+  # time in secs for which the URLs will be considered for fetching after a ack of fail
+  spout.ttl.purgatory: 30
+  
+  # Min time (in msecs) to allow between 2 successive queries to SOLR
+  spout.min.delay.queries: 2000
   
   # Solr MetricsConsumer
   solr.metrics.url: "http://localhost:8983/solr/metrics"
