+++ external/sql/src/main/java/com/digitalpebble/stormcrawler/sql/Constants.java
@@ -23,10 +23,16 @@ public class Constants {
     public static final String MYSQL_USER_PARAM_NAME = "mysql.user";
     public static final String MYSQL_PASSWORD_PARAM_NAME = "mysql.password";
     public static final String MYSQL_TABLE_PARAM_NAME = "mysql.table";
-    public static final String MYSQL_BUFFERSIZE_PARAM_NAME = "mysql.buffer.size";
     public static final String MYSQL_MAX_DOCS_BUCKET_PARAM_NAME = "mysql.max.urls.per.bucket";
-    public static final String MYSQL_MIN_QUERY_INTERVAL_PARAM_NAME = "mysql.min.query.interval";
     public static final String MYSQL_METRICS_TABLE_PARAM_NAME = "mysql.metrics.table";
+
+    public static final String MYSQL_MAXRESULTS_PARAM_NAME = "mysql.spout.max.results";
+
+    /**
+     * @deprecated replaced by
+     *             {@link com.digitalpebble.stormcrawler.persistence.AbstractQueryingSpout.StatusMinDelayParamName}
+     */
+    public static final String MYSQL_MIN_QUERY_INTERVAL_PARAM_NAME = "mysql.min.query.interval";
 
     private Constants() {
     }
